--- a/app/src/main/java/org/wikipedia/util/ActiveTimer.java
+++ b/app/src/main/java/org/wikipedia/util/ActiveTimer.java
@@ -26,4 +26,8 @@ public class ActiveTimer {
     public int getElapsedSec() {
         return (int) TimeUnit.MILLISECONDS.toSeconds(System.currentTimeMillis() - startMillis);
     }
+
+    public long getElapsedMillis() {
+        return System.currentTimeMillis() - startMillis;
+    }
 }