Comparison of load time

Old measurements taken from PR by @joewalsh : https://github.com/wikimedia/apps-android-wikipedia/commit/77861cf6f8148c906ec46b926c75387b8d3eee14

|Detail|Old Performance|mobile-html Performance|
|-------|-----------------|-------------|     
Page with image|~200ms|~600ms|
Page without image|~140ms|~400ms|
Re-open same page|~200ms|~150ms|
From existing tab|~160ms|~200ms|
 
**Note**: The page is painted a split second later during old measurements and mobile-html is fully painted