@joewalsh 2019-08-09
> Note: The page is painted a split second later during old measurements and mobile-html is fully painted 

I noticed this as well - feel free to adjust the measurement on my branch to reflect when the actual paint occurs.